BARK-113: Move user-models.ts into bark/models
